pyproject.toml
[build-system]
requires = [
    "hatchling>=1.10.0",
    "jupyterlab>=4.2.0,<5",
]
build-backend = "hatchling.build"

[project]
name = "bytegrader"
version = "0.0.1"
description = "Fork of Jupyter's nbgrader with full LTI support and optimized for JupyterHub with multiple courses and students"
#readme = "README.md"
requires-python = ">=3.10"
keywords = [
    "Grading",
    "Homework",
    "Jupyter",
    "JupyterLab",
    "Notebooks",
]
classifiers = [
    "Framework :: Jupyter",
    "Framework :: Jupyter :: JupyterLab",
    "Framework :: Jupyter :: JupyterLab :: 4",
    "Framework :: Jupyter :: JupyterLab :: Extensions",
    "Framework :: Jupyter :: JupyterLab :: Extensions :: Prebuilt",
    "License :: OSI Approved :: BSD License",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
]
dependencies = [
    "ipython>=8.10.0",
    "ipywidgets>=7.6",
    "Jinja2>=3",
    "jupyter_client<9",
    "jupyter_server>=2",
    "jupyterlab>=4.2.0,<5",
    "jupyterlab_server",
    "nbclient>=0.6.1",
    "nbconvert>=6",
    "notebook>=7.2.0,<8",
    "python-dateutil>=2.8",
    "rapidfuzz>=1.8",
    "requests>=2.26",
    "setuptools",
    "sqlalchemy>=1.4,<3",
    "PyYAML>=6.0",
    "PyJWT",
    "jupyterhub-systemdspawner>=1.0.2",
    "tomli>=2.2.1",
    "cachetools>=5.0.0",
    "cryptography>=3.4.8",
    "aiohttp>=3.8.0",
    "traitlets>=5.14.3",
    "rich>=14.0.0",
    "colorama>=0.4.6",
    "pydantic>=2.11.2",
    "apscheduler>=3.11.0",
]

[project.optional-dependencies]
wasm-executor = [
    "wasmtime>=35.0.0",
]

sentry = [
    "sentry-sdk>=2.0.0",
]

opentelemetry = [
    "opentelemetry-sdk>=1.26.0",
    "opentelemetry-exporter-otlp>=1.26.0",
    "opentelemetry-instrumentation-requests>=0.49b0",
    "opentelemetry-instrumentation-tornado>=0.49b0",
    "opentelemetry-instrumentation-sqlalchemy>=0.49b0",
]

[project.scripts]
bytegrader = "bytegrader.cli:main"

[project.entry-points."jupyter_serverproxy_servers"]
bytegrader = "bytegrader.extensions.hub:setup_server_proxy"

[tool.hatch.build]
artifacts = [
    "bytegrader/extensions/labextension/",
]

[tool.hatch.build.targets.sdist]
include = [
    "/etc",
    "/bytegrader",
    "/src",
    "/style",
    "/*.*"
]
exclude = [
    "/.github",
]

[tool.hatch.build.targets.wheel.shared-data]
"etc/jupyter/jupyter_server_config.d" = "etc/jupyter/jupyter_server_config.d"
"bytegrader/extensions/labextension" = "share/jupyter/labextensions/@bytechallenge/bytegrader"

[tool.hatch.build.hooks.jupyter-builder]
dependencies = [
    "hatch-jupyter-builder>=0.7",
]
build-function = "hatch_jupyter_builder.npm_builder"
ensured-targets = [
    "bytegrader/extensions/labextension/static/style.js",
    "bytegrader/extensions/labextension/package.json",
]
skip-if-exists = [
    "bytegrader/extensions/labextension/static/style.js",
]

[tool.hatch.build.hooks.jupyter-builder.editable-build-kwargs]
build_dir = "bytegrader/extensions/labextension"
source_dir = "src"
build_cmd = "install:labextension"
npm = [
    "jlpm",
]

[tool.hatch.build.hooks.jupyter-builder.build-kwargs]
build_cmd = "build:prod"
npm = [
    "jlpm",
]

[tool.jupyter-releaser.hooks]
after-bump-version = [
    "python tools/post-bump.py",
]
before-build-npm = [
    "python -m pip install 'jupyterlab>=4'",
    "jlpm",
    "jlpm build:prod",
]
before-build-python = [
    "jlpm clean:all",
]

[tool.tbump.version]
current = "0.9.5"
regex = '''
  (?P<major>\d+)\.(?P<minor>\d+)\.(?P<patch>\d+)
  ((?P<channel>a|b|rc|.dev)(?P<release>\d+))?
'''

[tool.tbump.git]
message_template = "Bump to {new_version}"
tag_template = "v{new_version}"

[[tool.tbump.before_commit]]
name = "check changelog"
cmd = "grep -q {new_version} CHANGELOG.md"

[[tool.tbump.file]]
src = "nbgrader/_version.py"
version_template = '({major}, {minor}, {patch}, "{channel}", "{release}")'

[[tool.tbump.file]]
src = "pyproject.toml"

[[tool.tbump.file]]
src = "src/assignment_list/index.ts"

[[tool.tbump.field]]
name = "channel"
default = ""

[[tool.tbump.field]]
name = "release"
default = ""

[tool.check-wheel-contents]
ignore = ["W002", "W004"]

[dependency-groups]
dev = [
    "freezegun>=1.5.4",
    "pytest>=8.4.1",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = [
    "unit: Unit tests (fast, isolated, mocked dependencies)",
    "integration: Integration tests (real database, limited external services)",
    "contract: Contract/API tests (schema validation, endpoint behavior)",
    "api: API/contract tests (HTTP endpoints, request/response schemas)",
    "security: Security-focused tests (auth, validation, injection)",
    "performance: Performance and load tests (may be slow)",
    "slow: Tests that take longer than usual to run",
    "auth: Tests requiring authentication setup",
    "database: Tests requiring database setup",
    "asyncio: Async tests requiring asyncio event loop",
]
